backend/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
pytest
httpx
supabase
PyJWT
